requirements.txt
opencv-python==4.8.1.78
numpy==1.26.4
imutils==0.5.4
matplotlib==3.8.2
python-vlc==3.0.18122
